Relocate and consolidate translation test data
