+++ server/service/src/sync/test/test_data/master_list.rs
@@ -1,0 +1,69 @@
+use repository::MasterListRow;
+
+use crate::sync::{
+    test::TestSyncPullRecord,
+    translations::{LegacyTableName, PullUpsertRecord},
+};
+
+const MASTER_LIST_1: (&'static str, &'static str) = (
+    "87027C44835B48E6989376F42A58F7EA",
+    r#"{
+    "ID": "87027C44835B48E6989376F42A58F7EA",
+    "description": "District Store",
+    "date_created": "2017-08-17",
+    "created_by_user_ID": "0763E2E3053D4C478E1E6B6B03FEC207",
+    "note": "note 1",
+    "gets_new_items": false,
+    "tags": null,
+    "isProgram": false,
+    "programSettings": null,
+    "code": "",
+    "isPatientList": false,
+    "is_hiv": false,
+    "isSupplierHubCatalog": false
+}"#,
+);
+
+const MASTER_LIST_2: (&'static str, &'static str) = (
+    "87027C44835B48E6989376F42A58F7E3",
+    r#"{
+    "ID": "87027C44835B48E6989376F42A58F7E3",
+    "description": "District Store 2",
+    "date_created": "2017-08-17",
+    "created_by_user_ID": "0763E2E3053D4C478E1E6B6B03FEC207",
+    "note": "note 2",
+    "gets_new_items": false,
+    "tags": null,
+    "isProgram": false,
+    "programSettings": null,
+    "code": "",
+    "isPatientList": false,
+    "is_hiv": false,
+    "isSupplierHubCatalog": false
+}"#,
+);
+
+pub(crate) fn test_pull_records() -> Vec<TestSyncPullRecord> {
+    vec![
+        TestSyncPullRecord::new_pull_upsert(
+            LegacyTableName::LIST_MASTER,
+            MASTER_LIST_1,
+            PullUpsertRecord::MasterList(MasterListRow {
+                id: MASTER_LIST_1.0.to_owned(),
+                name: "District Store".to_owned(),
+                code: "".to_owned(),
+                description: "note 1".to_owned(),
+            }),
+        ),
+        TestSyncPullRecord::new_pull_upsert(
+            LegacyTableName::LIST_MASTER,
+            MASTER_LIST_2,
+            PullUpsertRecord::MasterList(MasterListRow {
+                id: MASTER_LIST_2.0.to_owned(),
+                name: "District Store 2".to_owned(),
+                code: "".to_owned(),
+                description: "note 2".to_owned(),
+            }),
+        ),
+    ]
+}
